+++ src/zepben/evolve/services/network/tracing/tree/downstream_tree.py
@@ -5,13 +5,14 @@
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Optional, List, Set
+from typing import TYPE_CHECKING, Optional, Set
 
 from zepben.evolve.services.network.network_service import connected_terminals
 from zepben.evolve.exceptions import TracingException
 from zepben.evolve.services.network.tracing.feeder.feeder_direction import FeederDirection
 from zepben.evolve.services.network.tracing.traversals.queue import PriorityQueue
 from zepben.evolve.services.network.tracing.traversals.branch_recursive_tracing import BranchRecursiveTraversal
+from zepben.evolve.services.network.tracing.tree.tree_node import TreeNode
 from zepben.evolve.services.network.tracing.tree.tree_node_tracker import TreeNodeTracker
 if TYPE_CHECKING:
     from zepben.evolve import ConductingEquipment, Terminal, SinglePhaseKind
@@ -83,28 +84,3 @@ class DownstreamTree(object):
             raise TracingException(f"Missing conducting equipment for terminal {terminal.mrid}.")
 
         return set(filter(lambda phase: not self._open_test(conducting_equipment, phase), terminal.phases.single_phases))
-
-
-class TreeNode(object):
-
-    def __init__(self, conducting_equipment: ConductingEquipment, parent: Optional[TreeNode]):
-        self.conducting_equipment = conducting_equipment
-        self.parent = parent
-        self._children: List[TreeNode] = []
-        self._sort_weight = max((len(term.phases.single_phases) for term in conducting_equipment.terminals), default=1)
-
-    def __lt__(self, other: TreeNode):
-        """
-        This definition should only be used for sorting within a `PriorityQueue`
-
-        @param other: Another PhaseStep to compare against
-        @return: True if this node's max phase count over its equipment's terminals is greater than the other's, False otherwise.
-        """
-        return self._sort_weight > other._sort_weight
-
-    def __str__(self):
-        return f"{{conducting_equipment: {self.conducting_equipment.mrid}, parent: {self.parent and self.parent.conducting_equipment.mrid}, " \
-               f"num children: {len(self._children)}}}"
-
-    def add_child(self, child: TreeNode):
-        self._children.append(child)